src/python/requirements.txt
azure-iot-device==2.14.0
certifi==2025.6.15
charset-normalizer==3.4.2
deprecation==2.1.0
openai>=1.0.0
python-dotenv>=0.19.0
future==1.0.0
idna==3.10
iso8601==2.1.0
janus==2.0.0
packaging==25.0
paho-mqtt==1.6.1
pyserial==3.5
PySocks==1.7.1
PyYAML==6.0.2
requests==2.32.4
requests-unixsocket2==1.0.0
typing_extensions==4.14.1
urllib3==2.5.0
